Move tests to webhooks folder
